Creación del mapa
Se hizo que los tipos de chunk creados en el overworld dependan del valor correspondiente a su posición en el noise map

+++ core/Entities/components/TimedComponent.java
@@ -6,7 +6,7 @@ import com.badlogic.ashley.core.Component;
 
 public class TimedComponent implements Component{
 	
-	public long nextTurn = 0;
+	public int nextTurn = 0;
 	
 	public boolean isActive = true;
